backend/src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astroid==3.3.8
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dill==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fastapi==0.115.10
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
jwt==1.3.1
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pylint==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
